+++ tools/test-js-with-node
@@ -220,6 +220,7 @@ EXEMPT_FILES = make_set(
         "web/src/user_settings.ts",
         "web/src/user_status.js",
         "web/src/user_status_ui.js",
+        "web/src/user_topics.js",
         "web/src/webpack_public_path.js",
         "web/src/zcommand.js",
         "web/src/zform.js",
