+++ web/src/user_topics.js
@@ -1,5 +1,10 @@
+import render_topic_muted from "../templates/topic_muted.hbs";
+
 import * as blueslip from "./blueslip";
+import * as channel from "./channel";
+import * as feedback_widget from "./feedback_widget";
 import {FoldDict} from "./fold_dict";
+import {$t} from "./i18n";
 import {page_params} from "./page_params";
 import * as stream_data from "./stream_data";
 import * as timerender from "./timerender";
@@ -56,6 +61,54 @@ export function get_muted_topics() {
         }
     }
     return topics;
+}
+
+export function set_user_topic_visibility_policy(stream_id, topic, visibility_policy, from_hotkey) {
+    const data = {
+        stream_id,
+        topic,
+        visibility_policy,
+    };
+
+    channel.post({
+        url: "/json/user_topics",
+        data,
+        success() {
+            if (visibility_policy === all_visibility_policies.INHERIT) {
+                feedback_widget.dismiss();
+                return;
+            }
+            if (!from_hotkey) {
+                return;
+            }
+
+            // The following feedback_widget notice helps avoid
+            // confusion when a user who is not familiar with Zulip's
+            // keyboard UI hits "M" in the wrong context and has a
+            // bunch of messages suddenly disappear. This notice is
+            // only useful when muting from the keyboard, since you
+            // know what you did if you triggered muting with the
+            // mouse.
+            const stream_name = stream_data.maybe_get_stream_name(stream_id);
+            feedback_widget.show({
+                populate($container) {
+                    const rendered_html = render_topic_muted();
+                    $container.html(rendered_html);
+                    $container.find(".stream").text(stream_name);
+                    $container.find(".topic").text(topic);
+                },
+                on_undo() {
+                    set_user_topic_visibility_policy(
+                        stream_id,
+                        topic,
+                        all_visibility_policies.INHERIT,
+                    );
+                },
+                title_text: $t({defaultMessage: "Topic muted"}),
+                undo_button_text: $t({defaultMessage: "Undo mute"}),
+            });
+        },
+    });
 }
 
 export function set_user_topic(user_topic) {
