+++ web/tests/settings_muted_topics.test.js
@@ -7,7 +7,6 @@ const {run_test} = require("./lib/test");
 const $ = require("./lib/zjquery");
 
 const list_widget = mock_esm("../src/list_widget");
-const muted_topics_ui = mock_esm("../src/muted_topics_ui");
 
 const settings_muted_topics = zrequire("settings_muted_topics");
 const stream_data = zrequire("stream_data");
@@ -73,7 +72,7 @@ run_test("settings", ({override}) => {
     };
 
     let unmute_topic_called = false;
-    muted_topics_ui.unmute_topic = (stream_id, topic) => {
+    user_topics.set_user_topic_visibility_policy = (stream_id, topic) => {
         assert.equal(stream_id, frontend.stream_id);
         assert.equal(topic, "js");
         unmute_topic_called = true;
